+++ custom-sub-import.py
@@ -1,16 +1,18 @@
 #!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 
 """
 import abc
 import logging
 import os
+import re
 import shutil
 import sys
+from collections.abc import Callable
 from pathlib import Path
 from typing import Final
 
+from locate import LocatedSubtitles
 from locate import locate_english_subs_by_size
 from locate import locate_english_subs_vtx
 
@@ -31,13 +33,13 @@ class TestEventHandler(_EventHandler):
         self._log.info(f"Handling {event_type}")
 
 
-class _SubtitleCopier(object):
+class _SubtitleCopier:
     def __init__(
         self,
         log: logging.Logger,
         src_folder: Path,
         dest_folder: Path,
-        srt_locater,
+        srt_locater: Callable[[logging.Logger, Path], LocatedSubtitles],
     ):
         self._log = log
         self.src_folder = src_folder
@@ -72,32 +74,26 @@ class RadarrDownloadEventHandler(_EventHandler):
             os.environ.get("radarr_moviefile_sourcefolder"),
         )
 
+    def handle(self, event_type: str):
+        super().handle(event_type)
         # This is the destination folder Radarr will copy the movie file to
         media_destination_path = Path(os.environ.get("radarr_moviefile_path"))
 
         # The expected subtitle path is a simple Subs folder at the root level
         expected_subs_folder = self.download_file_src_folder / Path("Subs")
 
-        copier = None
+        # Assume the copy by size, then specialize if possible
+        copy_function = locate_english_subs_by_size
 
-        if "rarbg" in self.download_file_src_folder.name.lower():
+        if "vxt" in self.download_file_src_folder.name.lower():
+            copy_function = locate_english_subs_vtx
 
-            copier = _SubtitleCopier(
-                self._log,
-                expected_subs_folder,
-                media_destination_path,
-                locate_english_subs_by_size,
-            )
-        elif "vxt" in self.download_file_src_folder.name.lower():
-            copier = _SubtitleCopier(
-                self._log,
-                expected_subs_folder,
-                media_destination_path,
-                locate_english_subs_vtx,
-            )
-
-        if copier is not None:
-            copier.copy()
+        _SubtitleCopier(
+            self._log,
+            expected_subs_folder,
+            media_destination_path,
+            copy_function,
+        ).copy()
 
 
 class SonarrDownloadEventHandler(_EventHandler):
@@ -109,34 +105,66 @@ class SonarrDownloadEventHandler(_EventHandler):
             os.environ["sonarr_episodefile_sourcefolder"],
         )
 
+    def handle(self, event_type: str):
+        super().handle(event_type)
         media_destination_path = self.episode_file_path
 
-        copier = None
+        # The expected subtitle folder is Subs at the root level
+        base_subs_folder = self.episode_file_src_folder / Path("Subs")
 
-        if "rarbg" in self.episode_file_src_folder.name.lower():
+        if not base_subs_folder.exists():
+            self._log.error(
+                f"Subs folder {base_subs_folder} does not exist, can't do anything",
+            )
+            return
 
-            # The expected subtitle folder is Subs at the root level, with individual
-            # episode's subtitles in folders under that, named by the episode file name
-            expected_subs_folder = (
-                self.episode_file_src_folder
-                / Path("Subs")
-                / Path(self.episode_file_src_path.with_suffix("").name)
+        # Check against the episode name, minus extension
+        # episode's subtitles in folders under that, named by the episode file name
+        expected_subs_folder = base_subs_folder / Path(
+            self.episode_file_src_path.with_suffix("").name,
+        )
+
+        # Not found, try to locate by SxxEyy naming
+        if not expected_subs_folder.exists():
+            season_episode_re = re.compile("[Ss]\\d+[Ee]\\d+")
+            expected_subs_folder = None
+
+            # Find directories under Subs which appear to match SxxEyy naming
+            possible_sub_dirs = {}
+            for thing in base_subs_folder.glob("*"):
+                if not thing.is_dir():
+                    continue
+                match = season_episode_re.search(thing.name)
+                if match is not None:
+                    possible_sub_dirs[match.group(0)] = thing
+
+            match = season_episode_re.search(self.episode_file_src_path.name)
+            if match is not None:
+                season_episode = match.group(0)
+                if season_episode in possible_sub_dirs:
+                    expected_subs_folder = possible_sub_dirs[season_episode]
+
+        if expected_subs_folder is None:
+            self._log.error("Unable to locate subtitles folder")
+            return
+        else:
+            self._log.info(
+                f"Using {expected_subs_folder.relative_to(self.episode_file_src_path.parent)} "
+                f"for episode {self.episode_file_src_path.name}",
             )
 
-            copier = _SubtitleCopier(
-                self._log,
-                expected_subs_folder,
-                media_destination_path,
-                locate_english_subs_by_size,
-            )
+        copy_func = locate_english_subs_by_size
 
-        if copier is not None:
-            copier.copy()
+        _SubtitleCopier(
+            self._log,
+            expected_subs_folder,
+            media_destination_path,
+            copy_func,
+        ).copy()
 
 
 if __name__ == "__main__":
-
-    version: Final[str] = "0.1.0"
+    version: Final[str] = "0.2.0"
 
     main_log = logging.getLogger("custom-sub-import")
 
@@ -171,8 +199,9 @@ if __name__ == "__main__":
 
     handler_mapping = None
 
-    if "radarr_eventtype" in os.environ:
+    env_event_type = None
 
+    if "radarr_eventtype" in os.environ:
         env_event_type = os.environ["radarr_eventtype"]
 
         handler_mapping = radarr_events_to_handlers
@@ -182,7 +211,7 @@ if __name__ == "__main__":
 
         handler_mapping = sonarr_events_to_handlers
 
-    if env_event_type in handler_mapping:
+    if env_event_type is not None and env_event_type in handler_mapping:
         handler_mapping[env_event_type](main_log).handle(env_event_type)
     else:
         main_log.warning(f"Unhandled event: {env_event_type}")